fix delete tenant

+++ docs/examples/fastapi-integration/app/main.py
@@ -1,29 +1,17 @@
 import logging
-from typing import Sequence
 from uuid import UUID, uuid4
 
 from fastapi import FastAPI
 from pydantic import BaseModel
-from sqlalchemy import select
 from starlette import status
 from starlette.responses import JSONResponse, Response
 
 from app import orm
-from app.dependencies import Database_T
 from app.engine import manager
 
 logger = logging.getLogger(__name__)
 
 app = FastAPI()
-
-
-class TodoItemResp(BaseModel):
-    id: UUID
-    name: str
-
-
-class CreateTodoItemReq(BaseModel):
-    name: str
 
 
 class CreateTenantReq(BaseModel):
@@ -35,26 +23,6 @@ class CreateTenantResp(BaseModel):
     id: UUID
     slug: str
     description: str
-
-
-@app.get("/todos")
-async def list_todos(db: Database_T) -> Sequence[TodoItemResp]:
-    query = select(orm.TodoItem).where(orm.TodoItem.tenant == db.tenant)
-    result = await db.execute(query)
-    todos = result.scalars().all()
-    return [TodoItemResp(id=todo.id, name=todo.name) for todo in todos]
-
-
-@app.post("/todos")
-async def create_todo(db: Database_T, req: CreateTodoItemReq) -> TodoItemResp:
-    new_todo = orm.TodoItem(
-        id=uuid4(),
-        name=req.name,
-        tenant=UUID(str(db.tenant)),
-    )
-    db.add(new_todo)
-    await db.commit()
-    return TodoItemResp(id=new_todo.id, name=new_todo.name)
 
 
 @app.post("/tenants")
@@ -80,6 +48,13 @@ async def delete_tenant(
     tenant_id: UUID,
 ) -> Response:
     async with manager.new_session() as sess:
+        tenant = await sess.get(orm.Tenant, tenant_id)  # Ensure tenant exists
+        if tenant is None:
+            return JSONResponse(
+                content={"detail": f"tenant {tenant_id} not found."},
+                status_code=status.HTTP_404_NOT_FOUND,
+            )
+        await sess.delete(tenant)
         await manager.delete_tenant(tenant_id)
         await sess.commit()
         return JSONResponse(
